StyledCheckbox - Move to deprecated status. - Checkbox - Move Story to Inputs section.
