+++ src/deprecated/styled-checkbox.js
