+++ src/index.js
@@ -13,7 +13,6 @@ export { default as List } from './list'
 export { default as Loader } from './loader'
 export { default as QuickFilters } from './quick-filters'
 export { default as StyledCardContainer } from './styled-card-container'
-export { default as StyledCheckbox } from './styled-checkbox'
 export { default as StyledRadio } from './styled-radio'
 export { default as StyledSwitch } from './styled-switch'
 export { default as Table } from './table'
@@ -25,5 +24,6 @@ export { default as WidgetStats } from './widget-stats'
 // deprecated
 export { default as DataTable } from './deprecated/data-table' // deprecated by Table
 export { default as DynamicDataTable } from './deprecated/dynamic-data-table' // deprecated by Table
+export { default as StyledCheckbox } from './deprecated/styled-checkbox' // deprecated by Checkbox
 export { default as WidgetNumber } from './deprecated/widget-number' // deprecated by WidgetStats
 export { default as WidgetTrend } from './deprecated/widget-trend' // deprecated by WidgetStats
